+++ src/symbol/symbol-parse.test.js
@@ -1,13 +1,26 @@
 
 import { parse } from './symbol-parse';
 
-it('should parse symbol strings', () => {
-  expect(parse('-')).toEqual({
-  });
+it('should parse symbol strings for base', () => {
   expect(parse('𝠀')).toEqual({
     'base': '𝠀'
   });
 })
+
+
+it('should parse symbol strings', () => {
+  expect(parse('𝠀𝪛')).toEqual({'base': '𝠀','fill': '𝪛'});
+})
+
+it('should parse symbol strings', () => {
+  expect(parse('𝠀𝪡')).toEqual({'base': '𝠀','rotation':'𝪡'});
+})
+
+it('should parse symbol strings for base with fill and rotation', () => {
+  expect(parse('𝠀𝪛𝪡')).toEqual({'base': '𝠀','fill': '𝪛','rotation':'𝪡'});
+})
+
+
 
 it('should parse without breaking on bad data', () => {
   expect(parse()).toEqual({